split up repeats
