+++ src/iomusxml.cpp
@@ -35,13 +35,13 @@
 #include "measure.h"
 #include "mordent.h"
 #include "mrest.h"
+#include "mrpt.h"
 #include "note.h"
 #include "octave.h"
 #include "pb.h"
 #include "pedal.h"
 #include "rend.h"
 #include "rest.h"
-#include "rpt.h"
 #include "sb.h"
 #include "score.h"
 #include "section.h"
