+++ src/layer.cpp
@@ -26,8 +26,8 @@
 #include "measure.h"
 #include "mensur.h"
 #include "metersig.h"
+#include "mrpt.h"
 #include "note.h"
-#include "rpt.h"
 #include "staff.h"
 #include "staffdef.h"
 #include "vrv.h"
